+++ src/components/Modal/MyPageModal.tsx
@@ -1,11 +1,10 @@
 import CloseIcon from "@/assets/svgs/btnXbig.svg";
 import MyPage from "@/src/components/MyPage";
-import useModalStore from "@/src/stores/ModalStore";
+import useModalStore from "@/src/stores/modalStore";
 import { MyPageProps } from "@/src/types/user";
 
 const MyPageModal = ({ handleSubmit, isPending }: MyPageProps) => {
   const { setModalClose } = useModalStore();
-
 
   return (
     <div className="z-50 flex size-full items-center justify-center">
