+++ src/pages/my-page/index.tsx
@@ -7,7 +7,7 @@ import ProfileUpload from "@/src/components/SideNavigationMenu/common/ProfileUpl
 import PATH_NAMES from "@/src/constants/pathname";
 import useHydration from "@/src/hooks/useHydration";
 import { useUpdateMyData } from "@/src/queries/auth";
-import useModalStore from "@/src/stores/ModalStore";
+import useModalStore from "@/src/stores/modalStore";
 import useProfileImageUrlStore from "@/src/stores/useProfileImageUrlStore";
 import useUserStore from "@/src/stores/userStore";
 import { UserPatchProps } from "@/src/types/user";
